pyproject.toml
[project]
name = "agent-platform"
version = "0.1.0"
description = "FastAPI backend for unified agent orchestration"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache-2.0" }
keywords = ["fastapi", "agents", "rag"]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.31.0",
    "pydantic-settings>=2.6.0",
    "redis>=5.0.8",
    "supabase>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.3.0",
]

[build-system]
requires = ["setuptools>=68.0.0"]
build-backend = "setuptools.build_meta"
